refactor: split partial yaml loading into a separate module for re-use
