+++ tests/test_templates.py
@@ -12,12 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import stat
 import sys
 from pathlib import Path
 
-from synthtool.gcp import common
 from synthtool.sources import templates
 
 
@@ -106,21 +104,6 @@ def test_hide_billing():
         "README.md", metadata={"repo": {"requires_billing": False}}
     ).read_text()
     assert "Enable billing for your project" not in result
-
-
-def test_readme_partials():
-    cwd = os.getcwd()
-    os.chdir(FIXTURES)
-
-    common_templates = common.CommonTemplates()
-    metadata = {}
-    common_templates._load_partials(metadata)
-    # should have populated introduction from partial.
-    assert (
-        "objects to users via direct download" in metadata["partials"]["introduction"]
-    )
-
-    os.chdir(cwd)
 
 
 def test_ruby_authentication():
